# metal: minor q4 optimization and reduce code size