The first commit uses uint16_t instead of uint8_t in q4_0 and q4_1 kernels. Apple GPUs don't like 8-bit types, and for any operation on a 8-bit number the GPU will first copy it to an empty register and then do the calculation. This brings ~3% improvement for 33B model on M1 Max.<br>
 After this commit, we can achieve 340-350 GB/s memory read speed on M1 Max for 33B model, if we only run the q4_0-f32 MAT_MUL kernel and skip all other operations. This is very close to the reported hardware limit.
<img width="795" alt="Screenshot 2023-07-17 at 01 18 29" src="https://github.com/ggerganov/llama.cpp/assets/61452103/211fd36e-8509-43cb-9c16-ef485a2bc86f">
<br>(Only run matrix vector multiplications and skip all other operations.)

#
The second commit updates the RMS_NORM kernel by minimizing the use of threadgroup memory barrier, brings ~2% improvement for 33B model on M1 Max.

The third commit uses template to reduce code size. q5_0 and q5_1 support can be added quite efficiently using the new template. The new template also improve the behavior when `nb` is not divisible by 32. (thanks to discussion with @ikawrakow !)

Overall speed up (M1 Max, Updated as `f3f2e8e`):
||master| this PR|Speed up|
|---|---|---|---|
|33B q4_0  256 tokens| 82.1 ms/tok | 72.5 ms/tok| ~13%
|7B q4_0  256 tokens| 21.1 ms/tok | 19.6 ms/tok| ~7.6%

`./main -m model -n 256 -c 512 -s 123 -p "I believe the meaning of life is" -ngl 1 --no-mmap -t 8`