@lshzh-ww 2023-07-17
@ggerganov I plan to write a matrix-matrix multiplication kernel for metal backend, so that we can finally run prompt evaluation on GPU. I know that you want to avoid having different kernels for matrix-matrix and matrix-vector, but I found it's a bit hard to achieve high performance on both scenario with one kernel. 

This is part of the reason why I came up with this template. After this I can prepare a new template for q4_0, q4_1, q5_0 and q5_1 matrix-matrix multiplication. I guess we still need two more templates for q_k quantizations (they have larger block size that can't be put in registers, so the logic will be a little different). Overall there may be 4 templates and we can still have a reasonable code size. 

Let me know your opinions.