@lshzh-ww 2023-07-17
@ikawrakow BTW, you are right on the prefetch test. Using the disassembler from [here](https://github.com/dougallj/applegpu), it looks like the prefetch codes are optimized out by the compiler. I was test multiple optimizations at that time and wrongly attributed the speed up to prefetch.