@ggerganov 2023-07-17
I observe similar perf gains on M1 Pro:

| model | master | PR |
| --- | --- | --- |
| 7B  Q4_0 | 32.4 ms/t| 31.1 ms/t |
| 7B  Q4_1 | 34.6 ms/t| 33.5 ms/t |
| 13B  Q4_0 | 56.5 ms/t| 54.6 ms/t |
| 13B  Q4_1 | 60.9 ms/t| 59.4 ms/t |