@ggerganov 2023-07-17
> @ggerganov I plan to write a matrix-matrix multiplication kernel for metal backend, so that we can finally run prompt evaluation on GPU. I know that you want to avoid having different kernels for matrix-matrix and matrix-vector, but I found it's a bit hard to achieve high performance on both scenario with one kernel.

Please, go ahead. I haven't started working on this so I have no meaningful points to make. Reusing the dot-product code would be nice, but if it's not enough to obtain the best performance then we can do something else, for example what you propose.

There is some similar work going on in #2160 
Not sure if insights between both backends could be shared, but might be still useful to keep an eye

Btw, there is one strange thing that is observed with the Metal implementation, that would be nice to have an explanation of:

Do the following 2 runs:
- Put `return` on the first line of the matrix multiplication kernels
- Do not enqueue the matrix multiplications kernels at all

These 2 runs I would expect to have very similar performance, but for some reason the former one is noticeably slower.
At least this was the case last time I checked some time ago.

The difference was quite significant. I'm not a Metal expert, but it looked like a big cost for queuing an empty kernel.
Maybe worth looking into this and see if this overhead can be improved in some way. Might bring some extra performance gains