@lshzh-ww 2023-07-17
> The difference was quite significant. I'm not a Metal expert, but it looked like a big cost for queuing an empty kernel.

Yes there is latency even if you launch an empty kernel. I think that latency comes from that GPU has to prepare threadgroup memory and copy parameters to the constant buffer. On contemporary NVIDIA hardware people reported a kernel launch takes at least 5 us and starting every threadgroup takes ~0.5 us. Such number may be higher for metal.

There are quite a lot we can do to optimize performance, like making element-wise add/mul kernels use less threadgroups, removing unneeded parameters pass to kernels, etc. I am confident that we can speed up token generation by 10%-15% easily. 

Like you said some ideas from cuda backend may be useful, and I am keeping an eye on it.