@lshzh-ww 2023-07-20
The new commit revert the modification of the `block_q4_0` and `block_q4_0` (Thanks inspiration from @ikawrakow !). 

The new template also improve the behavior when `nb` is not divisible by 32: We first let each simd group calculate 32 blocks from the same row every time, until there are less than 32 blocks left for each row. In this case, we let each simd group loads 16 blocks from the first row and 16 blocks from the next row. By this design, we don't waste resource any more as long as `nb%16==0`.

The only exception is 7B model, in which part of the matrices have `ne00=11008` and `nb%32=24`. The last iteration for these matrices will have half of the threads idle. Hence the performance improvement for 7B won't be as big as others.